py-sqlalchemy-peewee-shared-db-conn/pyproject.toml
[tool.poetry]
name = "py-sqlalchemy-peewee-shared-db-conn"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.36"
peewee = "^3.17.8"
pytest = "^8.3.4"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
